From Master 7.5

+++ platforms/opencl/tests/TestOpenCLNonbondedForce.cpp
@@ -144,7 +144,7 @@ bool canRunHugeTest() {
 
     // Only run the huge test if the device has at least 4 GB of memory.
 
-    return (memory >= 4*(1<<30));
+    return (memory >= 4*(long long)(1<<30));
 }
 
 void runPlatformTests() {
